code refactoring and fix style
