+++ pam_sqlite3.c
@@ -1,58 +1,36 @@
 /*
- * PAM authentication module for SQLite
+ * pam_sqlite - sqlite database pam module
  *
- * SQLite3 port: Corey Henderson (cormander) <admin@ravencore.com>
- * SQLite port: Edin Kadribasic <edink@php.net>
- * Extended SQL configuration support by Wez Furlong <wez@thebrainroom.com>
+ * Copyright (C) 2022 A.Forouzan and A.Parker
  *
- * Based in part on pam_pgsql.c by David D.W. Downey ("pgpkeys") <david-downey@codecastle.com>
+ * This file is part of pam_sqlite
  *
- * Based in part on pam_unix.c of FreeBSD.
+ * pam_sqlite is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License v3 as published by the
+ * Free Software Foundation.
  *
-                          pam_sqlite3
-
-             This work is Copyright (C) Corey Henderson
-                           GNU GPL v3
- This work is a dirivative of pam_sqlite, Copyright (C) Edin Kadriba
-   pam_sqlite is a dirivative of libpam-pgsql, see the CREDITS file
-
-Different portions of this program are Copyright to the respective
-authors of those peices of code; but are all under the terms
-of of the GNU General Pulblic License.
-
-This program is free software; you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation; either version 2 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program; if not, write to the Free Software
-Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
-
+ * pam_sqlite is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; see the file COPYING; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-/* $Id: pam_sqlite.c,v 1.11 2003/07/17 13:47:07 wez Exp $ */
-
-#include "config.h"
 #include <stdio.h>
 #include <stdlib.h>
-#include <stdarg.h>
 #include <string.h>
-#include <syslog.h>
 #include <ctype.h>
+#include <sqlite3.h>
+
 #if HAVE_UNISTD_H
 #include <unistd.h>
 #endif
 #if HAVE_SYS_TYPES_H
 #include <sys/types.h>
 #endif
-#include <time.h>
-#include <sqlite3.h>
 #if HAVE_CRYPT_H
 #include <crypt.h>
 #endif
@@ -64,182 +42,18 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 #include <security/pam_appl.h>
 #include "pam_mod_misc.h"
 
+#include "config.h"
+#include "pam_sqlite3_password.h"
+#include "pam_sqlite3_conversation.h"
+#include "pam_sqlite3_db.h"
+#include "pam_sqlite3.h"
+
 #define PASSWORD_PROMPT			"Password: "
 #define PASSWORD_PROMPT_NEW		"New password: "
 #define PASSWORD_PROMPT_CONFIRM         "Confirm new password: "
-#define PASSWORD_TYPE_PROMT             "password type (1: plane, 2: MD5, 3: SHA256, 4: SHA512, 5: ENCRYPTED):"
-#define CONF				"/etc/pam_sqlite3.conf"
-
-#define DBGLOG(x...)  if(options->debug) {							\
-						  openlog("PAM_sqlite3", LOG_PID, LOG_AUTH); \
-						  syslog(LOG_DEBUG, ##x);					\
-						  closelog();								\
-					  }
-#define SYSLOG(x...)  do {											\
-						  openlog("PAM_sqlite3", LOG_PID, LOG_AUTH); \
-						  syslog(LOG_INFO, ##x);					\
-						  closelog();								\
-					  } while(0)
-#define SYSLOGERR(x...) SYSLOG("Error: " x)
-
-typedef enum {
-	PW_CLEAR = 1,
-#if HAVE_MD5_CRYPT
-	PW_MD5,
-#endif
-#if HAVE_SHA256_CRYPT
-	PW_SHA256,
-#endif
-#if HAVE_SHA512_CRYPT
-	PW_SHA512,
-#endif
-	PW_CRYPT,
-} pw_scheme;
-
-struct module_options {
-	char *database;
-	char *table;
-	char *user_column;
-	char *pwd_column;
-	char *pwd_type_column;
-	char *expired_column;
-	char *newtok_column;
-	pw_scheme default_pw_type;
-	int debug;
-	char *sql_verify;
-	char *sql_check_expired;
-	char *sql_check_newtok;
-	char *sql_set_passwd;
-};
-
-#define FAIL(MSG) 		\
-	{ 					\
-		SYSLOGERR(MSG);	\
-		free(buf); 		\
-		return NULL; 	\
-	}
-
-#define GROW(x)		if (x > buflen - dest - 1) {       		\
-	char *grow;                                        		\
-	buflen += 256 + x;                                 		\
-	grow = realloc(buf, buflen + 256 + x);             		\
-	if (grow == NULL) FAIL("Out of memory building query"); \
-	buf = grow;                                        		\
-}
-
-#define APPEND(str, len)	GROW(len); memcpy(buf + dest, str, len); dest += len
-#define APPENDS(str)	len = strlen(str); APPEND(str, len)
-
-#define MAX_ZSQL -1
-
-/*
- * Being very defensive here. The current logic in the rest of the code should prevent this from
- * happening. But lets protect against future code changes which could cause a NULL ptr to creep
- * in.
- */
-#define CHECK_STRING(str) 													 	\
-	if (!str) 															    	\
-		FAIL("Internal error in format_query: string ptr " #str " was NULL");
-
-static char *format_query(const char *template, struct module_options *options,
-	const char *user, const char *passwd)
-{
-	char *buf = malloc(256);
-	if (!buf)
-		return NULL;
-
-	int buflen = 256;
-	int dest = 0, len;
-	const char *src = template;
-	char *pct;
-	char *tmp;
-
-	while (*src) {
-		pct = strchr(src, '%');
-
-		if (pct) {
-			/* copy from current position to % char into buffer */
-			if (pct != src) {
-				len = pct - src;
-				APPEND(src, len);
-			}
-
-			/* decode the escape */
-			switch(pct[1]) {
-				case 'U':	/* username */
-					if (user) {
-						tmp = sqlite3_mprintf("%q", user);
-						if (!tmp)
-							FAIL("sqlite3_mprintf out of memory");
-						len = strlen(tmp);
-						APPEND(tmp, len);
-						sqlite3_free(tmp);
-					}
-					break;
-				case 'P':	/* password */
-					if (passwd) {
-						tmp = sqlite3_mprintf("%q", passwd);
-						if (!tmp)
-							FAIL("sqlite3_mprintf out of memory");
-						len = strlen(tmp);
-						APPEND(tmp, len);
-						sqlite3_free(tmp);
-					}
-					break;
-
-				case 'O':	/* option value */
-					pct++;
-					switch (pct[1]) {
-						case 'p':	/* passwd */
-							CHECK_STRING(options->pwd_column);
-							APPENDS(options->pwd_column);
-							break;
-						case 'k':	/* passwd type */
-						CHECK_STRING(options->pwd_type_column);
-							APPENDS(options->pwd_type_column);
-							break;
-						case 'u':	/* username */
-							CHECK_STRING(options->user_column);
-							APPENDS(options->user_column);
-							break;
-						case 't':	/* table */
-							CHECK_STRING(options->table);
-							APPENDS(options->table);
-							break;
-						case 'x':	/* expired */
-							CHECK_STRING(options->expired_column);
-							APPENDS(options->expired_column);
-							break;
-						case 'n':	/* newtok */
-							CHECK_STRING(options->newtok_column);
-							APPENDS(options->newtok_column);
-							break;
-					}
-					break;
-
-				case '%':	/* quoted % sign */
-					APPEND(pct, 1);
-					break;
-
-				default:	/* unknown */
-					APPEND(pct, 2);
-					break;
-			}
-			src = pct + 2;
-		} else {
-			/* copy rest of string into buffer and we're done */
-			len = strlen(src);
-			APPEND(src, len);
-			break;
-		}
-	}
-
-	buf[dest] = '\0';
-	return buf;
-}
-
-static void
-get_module_options_from_file(const char *filename, struct module_options *opts, int warn);
+#define PASSWORD_PROMPT_TYPE            "Password type (1: plane, 2: MD5, 3: SHA256, 4: SHA512, 5: ENCRYPTED):"
+#define CONFIG_FILE_PATH		"/etc/pam_sqlite3.conf"
+#define UNKNOWN_SERVICE                 "<Unknown Service>"
 
 /*
  * safe_assign protects against duplicate config options causing a memory leak.
@@ -251,6 +65,22 @@ safe_assign(char **asignee, const char *val)
 		free(*asignee);
 	*asignee = strdup(val);
 }
+
+void memzero_explicit(void *s, size_t cnt)
+{
+	memset(s, 0, cnt);
+	__asm__ __volatile__("": :"r"(s): "memory");
+}
+
+const char* pam_get_service(pam_handle_t *pamh, const char **service)
+{
+	if (pam_get_item(pamh, PAM_SERVICE, (const void**)service) != PAM_SUCCESS)
+		*service = UNKNOWN_SERVICE;
+	return *service;
+}
+
+static void
+get_module_options_from_file(const char *filename, struct module_options *opts, int warn);
 
 /* private: parse and set the specified string option */
 static void
@@ -293,31 +123,31 @@ set_module_option(const char *option, struct module_options *options)
 	} else if(!strcmp(buf, "user_column")) {
 		safe_assign(&options->user_column, val);
 	} else if(!strcmp(buf, "pwd_column")) {
-		safe_assign(&options->pwd_column, val);
+		safe_assign(&options->passwd_column, val);
 	} else if(!strcmp(buf, "pwd_type_column")) {
-		safe_assign(&options->pwd_type_column, val);
+		safe_assign(&options->passwd_type_column, val);
 	} else if(!strcmp(buf, "expired_column")) {
 		safe_assign(&options->expired_column, val);
 	} else if(!strcmp(buf, "newtok_column")) {
 		safe_assign(&options->newtok_column, val);
 	} else if(!strcmp(buf, "default_pw_type")) {
-		options->default_pw_type = PW_CLEAR;
+		options->default_passwd_type = PW_CLEAR;
 		if(!strcmp(val, "crypt")) {
-			options->default_pw_type = PW_CRYPT;
+			options->default_passwd_type = PW_CRYPT;
 		}
 #if HAVE_MD5_CRYPT
 		else if(!strcmp(val, "md5")) {
-			options->default_pw_type = PW_MD5;
+			options->default_passwd_type = PW_MD5;
 		}
 #endif
 #if HAVE_SHA256_CRYPT
 		else if(!strcmp(val, "sha-256")) {
-			options->default_pw_type = PW_SHA256;
+			options->default_passwd_type = PW_SHA256;
 		}
 #endif
 #if HAVE_SHA512_CRYPT
 		else if(!strcmp(val, "sha-512")) {
-			options->default_pw_type = PW_SHA512;
+			options->default_passwd_type = PW_SHA512;
 		}
 #endif
 	} else if(!strcmp(buf, "debug")) {
@@ -366,6 +196,37 @@ get_module_options_from_file(const char *filename, struct module_options *opts, 
 	}
 }
 
+/*
+ * If the given name is a standard option, set the corresponding flag in
+ * the options word and return 0.  Else return -1.
+ */
+static int
+pam_std_option(int *options, const char *name)
+{
+	struct opttab {
+		const char *name;
+		int value;
+	};
+	static struct opttab std_options[] = {
+		{ "debug",          PAM_OPT_DEBUG },
+		{ "no_warn",        PAM_OPT_NO_WARN },
+		{ "use_first_pass", PAM_OPT_USE_FIRST_PASS },
+		{ "try_first_pass", PAM_OPT_TRY_FIRST_PASS },
+		{ "use_mapped_pass",PAM_OPT_USE_MAPPED_PASS },
+		{ "echo_pass",      PAM_OPT_ECHO_PASS },
+		{ NULL,         0 }
+	};
+	struct opttab *p;
+
+	for (p = std_options;  p->name != NULL;  p++) {
+		if (strcmp(name, p->name) == 0) {
+			*options |= p->value;
+			return 0;
+		}
+	}
+	return -1;
+}
+
 /* private: read module options from file or commandline */
 static int
 get_module_options(int argc, const char **argv, struct module_options **options)
@@ -380,9 +241,9 @@ get_module_options(int argc, const char **argv, struct module_options **options)
 	}
 
 	bzero(opts, sizeof(*opts));
-	opts->default_pw_type = PW_CLEAR;
+	opts->default_passwd_type = PW_CLEAR;
 
-	get_module_options_from_file(CONF, opts, 0);
+	get_module_options_from_file(CONFIG_FILE_PATH, opts, 0);
 
 	for(i = 0; i < argc; i++) {
 		if(pam_std_option(&rc, argv[i]) == 0)
@@ -407,8 +268,8 @@ free_module_options(struct module_options *options)
 		free(options->table);
 	if(options->user_column)
 		free(options->user_column);
-	if(options->pwd_column)
-		free(options->pwd_column);
+	if(options->passwd_column)
+		free(options->passwd_column);
 	if(options->expired_column)
 		free(options->expired_column);
 	if(options->newtok_column)
@@ -444,206 +305,12 @@ options_valid(struct module_options *options)
 	return 0;
 }
 
-/* private: open SQLite database */
-static sqlite3 *pam_sqlite3_connect(struct module_options *options)
-{
-  const char *errtext = NULL;
-  sqlite3 *sdb = NULL;
-
-  if (sqlite3_open(options->database, &sdb) != SQLITE_OK) {
-      errtext = sqlite3_errmsg(sdb);
-	  SYSLOG("Error opening SQLite database (%s)", errtext);
-	  /*
-	   * N.B. sdb is usually non-NULL when errors occur, so we explicitly
-	   * release the resource and return NULL to indicate failure to the caller.
-	   */
-
-	  sqlite3_close(sdb);
-	  return NULL;
-  }
-
-  return sdb;
-}
-
-/* private: generate random salt character */
-static char *
-crypt_make_salt(pw_scheme pw_type)
-{
-	int i;
-	time_t now;
-	static unsigned long x;
-	static char result[22];
-	static char salt_chars[] = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789./";
-	static const int NUM_SALT_CHARS = sizeof(salt_chars) / sizeof(salt_chars[0]);
-
-	time(&now);
-	x += now + getpid() + clock();
-	srandom(x);
-
-	for (i=0; i<19; i++) {
-		result[i] = salt_chars[random() % NUM_SALT_CHARS];
-	}
-	result[i+1] = '$';
-	result[i+2]='\0';
-
-	switch(pw_type) {
-	case PW_CRYPT:
-		result[2] = '\0';
-		break;
-#if HAVE_MD5_CRYPT
-	case PW_MD5:
-		result[0]='$';
-		result[1]='1';
-		result[2]='$';
-		break;
-#endif
-#if HAVE_SHA256_CRYPT
-	case PW_SHA256:
-		result[0]='$';
-		result[1]='5';
-		result[2]='$';
-		break;
-#endif
-#if HAVE_SHA512_CRYPT
-	case PW_SHA512:
-		result[0]='$';
-		result[1]='6';
-		result[2]='$';
-		break;
-#endif
-	default:
-		result[0] = '\0';
-	}
-
-	return result;
-}
-
-/* private: encrypt password using the preferred encryption scheme */
-static char *
-encrypt_password(pw_scheme pw_type, const char *pass)
-{
-	char *s = NULL;
-
-	switch(pw_type) {
-#if HAVE_MD5_CRYPT
-		case PW_MD5:
-#endif
-#if HAVE_SHA256_CRYPT
-		case PW_SHA256:
-#endif
-#if HAVE_SHA512_CRYPT
-		case PW_SHA512:
-#endif
-		case PW_CRYPT:
-			s = strdup(crypt(pass, crypt_make_salt(pw_type)));
-			break;
-		case PW_CLEAR:
-		default:
-			s = strdup(pass);
-	}
-	return s;
-}
-
-/* private: authenticate user and passwd against database */
-static int
-auth_verify_password(const char *user, const char *passwd,
-					 struct module_options *options)
-{
-	int res;
-	pw_scheme pw_type = PW_CLEAR;
-	sqlite3 *conn = NULL;
-	sqlite3_stmt *vm = NULL;
-	int rc = PAM_AUTH_ERR;
-	const char *tail  = NULL;
-	const char *errtext = NULL;
-	const char *encrypted_pw = NULL;
-	char *query  = NULL;
-
-	if(!(conn = pam_sqlite3_connect(options))) {
-		rc = PAM_AUTH_ERR;
-		goto done;
-	}
-
-	if(!(query = format_query(options->sql_verify ? options->sql_verify :
-			"SELECT %Op, %Ok FROM %Ot WHERE %Ou='%U'",
-			options, user, passwd))) {
-		SYSLOGERR("failed to construct sql query");
-		rc = PAM_AUTH_ERR;
-		goto done;
-	}
-
-	DBGLOG("query: %s", query);
-
-	res = sqlite3_prepare(conn, query, MAX_ZSQL, &vm, &tail);
-
-	free(query);
-
-	if (res != SQLITE_OK) {
-        errtext = sqlite3_errmsg(conn);
-		DBGLOG("Error executing SQLite query (%s)", errtext);
-		rc = PAM_AUTH_ERR;
-		goto done;
-	}
-
-	if (SQLITE_ROW != sqlite3_step(vm)) {
-		rc = PAM_USER_UNKNOWN;
-		DBGLOG("no rows to retrieve");
-	} else {
-		const char *stored_pw = (const char *) sqlite3_column_text(vm, 0);
-		if (!stored_pw) {
-			SYSLOG("sqlite3 failed to return row data");
-			rc = PAM_AUTH_ERR;
-			goto done;
-		}
-
-		pw_type = sqlite3_column_int(vm, 1);
-		if (!pw_type)
-			pw_type = PW_CLEAR;
-
-		SYSLOG("pw type: %d", pw_type);
-
-		switch(pw_type) {
-#if HAVE_MD5_CRYPT
-		case PW_MD5:
-#endif
-#if HAVE_SHA256_CRYPT
-		case PW_SHA256:
-#endif
-#if HAVE_SHA512_CRYPT
-		case PW_SHA512:
-#endif
-		case PW_CRYPT:
-			encrypted_pw = crypt(passwd, stored_pw);
-			if (!encrypted_pw) {
-				SYSLOG("crypt failed when encrypting password");
-				rc = PAM_AUTH_ERR;
-				goto done;
-			}
-
-			if(strcmp(encrypted_pw, stored_pw) == 0)
-				rc = PAM_SUCCESS;
-			break;
-
-		/* PW_CLEAR */
-		default:
-			if(strcmp(passwd, stored_pw) == 0)
-				rc = PAM_SUCCESS;
-			break;
-		}
-	}
-
-done:
-	sqlite3_finalize(vm);
-	sqlite3_close(conn);
-	return rc;
-}
-
 /* public: authenticate user */
 PAM_EXTERN int
 pam_sm_authenticate(pam_handle_t *pamh, int flags, int argc, const char **argv)
 {
 	struct module_options *options = NULL;
-	const char *user = NULL, *password = NULL, *service = NULL;
+	const char *user = NULL, *passwd = NULL, *service = NULL;
 	int rc, std_flags;
 
 	std_flags = get_module_options(argc, argv, &options);
@@ -659,12 +326,12 @@ pam_sm_authenticate(pam_handle_t *pamh, int flags, int argc, const char **argv)
 
 	DBGLOG("attempting to authenticate: %s", user);
 
-	if((rc = pam_get_pass(pamh, &password, PASSWORD_PROMPT, std_flags)
+	if((rc = pam_get_pass(pamh, &passwd, PASSWORD_PROMPT, std_flags)
 		!= PAM_SUCCESS)) {
 		goto done;
 	}
 
-	if((rc = auth_verify_password(user, password, options)) != PAM_SUCCESS)
+	if((rc = auth_verify_passwd(user, passwd, options)) != PAM_SUCCESS)
 		SYSLOG("(%s) user %s not authenticated.", pam_get_service(pamh, &service), user);
 	else
 		SYSLOG("(%s) user %s authenticated.", pam_get_service(pamh, &service), user);
@@ -803,7 +470,7 @@ pam_sm_chauthtok(pam_handle_t *pamh, int flags, int argc, const char **argv)
 	char *errtext = NULL;
 	char *query = NULL;
 	int res;
-	pw_scheme pw_type;
+	passwd_scheme passwd_type;
 
 	std_flags = get_module_options(argc, argv, &options);
 	if(options_valid(options) != 0) {
@@ -819,7 +486,7 @@ pam_sm_chauthtok(pam_handle_t *pamh, int flags, int argc, const char **argv)
 	if(flags & PAM_PRELIM_CHECK) {
 		/* at this point, this is the first time we get called */
 		if((rc = pam_get_pass(pamh, &pass, PASSWORD_PROMPT, std_flags)) == PAM_SUCCESS) {
-			if((rc = auth_verify_password(user, pass, options)) == PAM_SUCCESS) {
+			if((rc = auth_verify_passwd(user, pass, options)) == PAM_SUCCESS) {
 				rc = pam_set_item(pamh, PAM_OLDAUTHTOK, (const void *)pass);
 				if(rc != PAM_SUCCESS) {
 					SYSLOGERR("failed to set PAM_OLDAUTHTOK!");
@@ -839,7 +506,7 @@ pam_sm_chauthtok(pam_handle_t *pamh, int flags, int argc, const char **argv)
 			SYSLOGERR("could not retrieve old token");
 			goto done;
 		}
-		rc = auth_verify_password(user, pass, options);
+		rc = auth_verify_passwd(user, pass, options);
 		if(rc != PAM_SUCCESS) {
 			SYSLOG("(%s) user '%s' not authenticated.", pam_get_service(pamh, &service), user);
 			goto done;
@@ -860,21 +527,21 @@ pam_sm_chauthtok(pam_handle_t *pamh, int flags, int argc, const char **argv)
 		}
 
 		/* get password type from user */
-		rc = pam_conversation(pamh, PASSWORD_TYPE_PROMT, std_flags, &pass_type);
+		rc = pam_conversation(pamh, PASSWORD_PROMPT_TYPE, std_flags, &pass_type);
 		if(rc != PAM_SUCCESS) {
 			SYSLOGERR("failed to get pass type!");
 			goto done;
 		}
 
-		pw_type = atoi(pass_type);
-		if(pw_type == 0) {
-			SYSLOGERR("invalid pw type str!");
+		passwd_type = atoi(pass_type);
+		if(passwd_type == 0) {
+			SYSLOGERR("invalid passwd type str!");
 			rc = PAM_SYMBOL_ERR;
 			goto done;
 		}
 
 		/* update the database */
-		if(!(newpass_crypt = encrypt_password(pw_type, newpass))) {
+		if(!(newpass_crypt = encrypt_passwd(passwd_type, newpass))) {
 			SYSLOGERR("passwd encrypt failed");
 			rc = PAM_BUF_ERR;
 			goto done;
@@ -929,12 +596,3 @@ pam_sm_setcred(pam_handle_t *pamh, int flags, int argc, const char **argv)
 {
 	return PAM_SUCCESS;
 }
-
-/*
- * Local variables:
- * tab-width: 4
- * c-basic-offset: 4
- * End:
- * vim600: sw=4 ts=4 fdm=marker
- * vim<600: sw=4 ts=4
- */
