+++ pam_sqlite3_password.h
@@ -19,17 +19,13 @@
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-#ifndef PAM_MOD_MISC_H
-#define PAM_MOD_MISC_H
+#ifndef PAM_SQLITE3_PAM_SQLITE3_PASSWORD_H
+#define PAM_SQLITE3_PAM_SQLITE3_PASSWORD_H
 
-#include <sys/cdefs.h>
+#include "pam_sqlite3.h"
 
-/* Options */
-#define PAM_OPT_DEBUG			0x01
-#define PAM_OPT_NO_WARN			0x02
-#define PAM_OPT_USE_FIRST_PASS		0x04
-#define	PAM_OPT_TRY_FIRST_PASS		0x08
-#define PAM_OPT_USE_MAPPED_PASS		0x10
-#define PAM_OPT_ECHO_PASS		0x20
+char * encrypt_passwd(passwd_scheme pw_type, const char *pass);
+int auth_verify_passwd(const char *username, const char *passwd,
+		       struct module_options *options);
 
-#endif
+#endif //PAM_SQLITE3_PAM_SQLITE3_PASSWORD_H
